AI_Employee/requirements.txt
watchdog
playwright
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client